requirements.txt
numpy==2.3.1
PyYAML==6.0.2
tqdm==4.67.1
